requirements.txt
alembic~=1.5.7
boto3==1.20.24
colorama~=0.4.4
deepdiff~=5.6.0
Deprecated~=1.2.12
emails~=0.6
email-validator~=1.1.3
fastapi~=0.63.0
fastapi-mail~=0.3.5.0
gunicorn~=20.0.4
itsdangerous==1.1.0
more-itertools~=8.7.0
passlib[bcrypt]==1.7.4
psycopg2-binary~=2.8.6
pydantic[email]~=1.9.0
pyjwt==2.1.0  # Todo: Not sure if we need this one
python-dotenv~=0.17.1
python-jose[cryptography]==3.2.0
python-rapidjson==1.0
pytz>=2019.1
SQLAlchemy~=1.3.24
SQLAlchemy-Searchable~=1.2.0
structlog~=22.1.0
typer-cli~=0.0.11
uvicorn[standard]~=0.19.0
mangum>=0.15.1
